Refactored snakesAndLAdders package names
