+++ GameData/playerData.csv
@@ -4,3 +4,5 @@ Luna,3
 Mio,4
 n,2
 p,2
+Joffen,4
+Roar,2
