+++ src/main/java/edu/ntnu/idi/idattx2002/gui/ladderGameGui/view/TilesWindow.java
@@ -1,12 +1,12 @@
-package edu.ntnu.idi.idattx2002.gui.view;
-import edu.ntnu.idi.idattx2002.logic.Board.Actions.LadderAction;
-import edu.ntnu.idi.idattx2002.logic.Board.Actions.SwitchWithRandomAction;
-import edu.ntnu.idi.idattx2002.logic.Board.Board;
-import edu.ntnu.idi.idattx2002.logic.Board.LandAction;
-import edu.ntnu.idi.idattx2002.logic.Board.Tile;
-import edu.ntnu.idi.idattx2002.logic.Games.SnakesAndLadders;
-import edu.ntnu.idi.idattx2002.logic.PlayerObserver;
-import edu.ntnu.idi.idattx2002.logic.Player.Player;
+package edu.ntnu.idi.idattx2002.gui.ladderGameGui.view;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Board.Actions.LadderAction;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Board.Actions.SwitchWithRandomAction;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Board.Board;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Board.LandAction;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Board.Tile;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Games.SnakesAndLadders;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.PlayerObserver;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Player.Player;
 import java.util.ArrayList;
 import java.util.Map;
 import javafx.animation.PauseTransition;
