--- src/main/java/edu/ntnu/idi/idattx2002/logic/PlayerObserver.java
@@ -1,9 +1,0 @@
-package edu.ntnu.idi.idattx2002.logic;
-
-import edu.ntnu.idi.idattx2002.logic.Player.Player;
-
-public interface PlayerObserver {
-
-  void update(Player player);
-
-}
