+++ src/main/java/edu/ntnu/idi/idattx2002/logic/ladderGameLogic/Board/LandAction.java
@@ -1,0 +1,7 @@
+package edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Board;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Player.Player;
+
+public interface LandAction {
+  void perform(Player player);
+}
+
