+++ src/main/java/edu/ntnu/idi/idattx2002/logic/ladderGameLogic/DiceObserver.java
@@ -1,0 +1,8 @@
+package edu.ntnu.idi.idattx2002.logic.ladderGameLogic;
+
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Dice.Dice;
+
+public interface DiceObserver {
+
+  public void update(Dice dice);
+}
