+++ src/main/java/edu/ntnu/idi/idattx2002/logic/ladderGameLogic/PlayerObserver.java
@@ -1,0 +1,9 @@
+package edu.ntnu.idi.idattx2002.logic.ladderGameLogic;
+
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Player.Player;
+
+public interface PlayerObserver {
+
+  void update(Player player);
+
+}
