+++ src/test/java/PlayerTest.java
@@ -1,4 +1,4 @@
-import edu.ntnu.idi.idattx2002.logic.Player.Player;
+import edu.ntnu.idi.idattx2002.logic.ladderGameLogic.Player.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
